python_flask_app_with_nginx_reverse_proxy/part3_flask_mysql_ngnix/requirements.txt
Flask==2.3.3
mysql-connector-python
gunicorn
